fix: functions
